Add documentation on generating page from a nested class

+++ documentationtesting/src/test/java/org/sfvl/test_tools/ProjectTestExtension.java
@@ -17,8 +17,11 @@ public class ProjectTestExtension extends SimpleApprovalsExtension {
     }
 
     public void removeNonApprovalFiles(DocPath docPath) {
-        for(File file: docPath.page().path().getParent().toFile().listFiles(pathname ->
-                pathname.isFile() && !pathname.getName().startsWith("_"))) {
+        final File[] filesToDelete = docPath.page().path().getParent().toFile().listFiles(
+                pathname -> pathname.isFile() && !pathname.getName().startsWith("_"));
+        if (filesToDelete == null) return;
+
+        for(File file: filesToDelete) {
             file.delete();
         }
     }
